+++ src/client.ts
@@ -1,22 +1,24 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import type { RequestInit, RequestInfo, BodyInit } from './internal/builtin-types';
-import type { HTTPMethod, PromiseOrValue, MergedRequestInit } from './internal/types';
+import type { HTTPMethod, PromiseOrValue, MergedRequestInit, FinalizedRequestInit } from './internal/types';
 import { uuid4 } from './internal/utils/uuid';
-import { validatePositiveInteger, isAbsoluteURL, hasOwn } from './internal/utils/values';
+import { validatePositiveInteger, isAbsoluteURL, safeJSON } from './internal/utils/values';
 import { sleep } from './internal/utils/sleep';
+import { type Logger, type LogLevel, parseLogLevel } from './internal/utils/log';
+export type { Logger, LogLevel } from './internal/utils/log';
 import { castToError, isAbortError } from './internal/errors';
 import type { APIResponseProps } from './internal/parse';
 import { getPlatformHeaders } from './internal/detect-platform';
 import * as Shims from './internal/shims';
 import * as Opts from './internal/request-options';
 import { VERSION } from './version';
-import * as Errors from './error';
-import * as Pagination from './pagination';
-import { AbstractPage, type EntriesCursorParams, EntriesCursorResponse } from './pagination';
-import * as Uploads from './uploads';
+import * as Errors from './core/error';
+import * as Pagination from './core/pagination';
+import { AbstractPage, type EntriesCursorParams, EntriesCursorResponse } from './core/pagination';
+import * as Uploads from './core/uploads';
 import * as API from './resources/index';
-import { APIPromise } from './api-promise';
+import { APIPromise } from './core/api-promise';
 import { type Fetch } from './internal/builtin-types';
 import { HeadersLike, NullableHeaders, buildHeaders } from './internal/headers';
 import { FinalRequestOptions, RequestOptions } from './internal/request-options';
@@ -146,48 +148,6 @@ import {
   WorkflowsEntriesCursor,
 } from './resources/workflows/workflows';
 
-const safeJSON = (text: string) => {
-  try {
-    return JSON.parse(text);
-  } catch (err) {
-    return undefined;
-  }
-};
-
-type LogFn = (message: string, ...rest: unknown[]) => void;
-export type Logger = {
-  error: LogFn;
-  warn: LogFn;
-  info: LogFn;
-  debug: LogFn;
-};
-export type LogLevel = 'off' | 'error' | 'warn' | 'info' | 'debug';
-const parseLogLevel = (
-  maybeLevel: string | undefined,
-  sourceName: string,
-  client: Knock,
-): LogLevel | undefined => {
-  if (!maybeLevel) {
-    return undefined;
-  }
-  const levels: Record<LogLevel, true> = {
-    off: true,
-    error: true,
-    warn: true,
-    info: true,
-    debug: true,
-  };
-  if (hasOwn(levels, maybeLevel)) {
-    return maybeLevel;
-  }
-  loggerFor(client).warn(
-    `${sourceName} was set to ${JSON.stringify(maybeLevel)}, expected one of ${JSON.stringify(
-      Object.keys(levels),
-    )}`,
-  );
-  return undefined;
-};
-
 export interface ClientOptions {
   /**
    * Defaults to process.env['KNOCK_SERVICE_TOKEN'].
@@ -197,7 +157,7 @@ export interface ClientOptions {
   /**
    * Override the default base URL for the API, e.g., "https://api.example.com/v2/"
    *
-   * Defaults to process.env['KNOCK_BASE_URL'].
+   * Defaults to process.env['KNOCK_MGMT_BASE_URL'].
    */
   baseURL?: string | null | undefined;
 
@@ -249,7 +209,7 @@ export interface ClientOptions {
   /**
    * Set the log level.
    *
-   * Defaults to process.env['KNOCK_LOG'] or 'warn' if it isn't set.
+   * Defaults to process.env['KNOCK_MGMT_LOG'] or 'warn' if it isn't set.
    */
   logLevel?: LogLevel | undefined;
 
@@ -261,12 +221,10 @@ export interface ClientOptions {
   logger?: Logger | undefined;
 }
 
-type FinalizedRequestInit = RequestInit & { headers: Headers };
-
 /**
- * API Client for interfacing with the Knock API.
+ * API Client for interfacing with the Knock Mgmt API.
  */
-export class Knock {
+export class KnockMgmt {
   serviceToken: string;
 
   baseURL: string;
@@ -282,10 +240,10 @@ export class Knock {
   private _options: ClientOptions;
 
   /**
-   * API Client for interfacing with the Knock API.
+   * API Client for interfacing with the Knock Mgmt API.
    *
    * @param {string | undefined} [opts.serviceToken=process.env['KNOCK_SERVICE_TOKEN'] ?? undefined]
-   * @param {string} [opts.baseURL=process.env['KNOCK_BASE_URL'] ?? https://control.knock.app] - Override the default base URL for the API.
+   * @param {string} [opts.baseURL=process.env['KNOCK_MGMT_BASE_URL'] ?? https://control.knock.app] - Override the default base URL for the API.
    * @param {number} [opts.timeout=1 minute] - The maximum amount of time (in milliseconds) the client will wait for a response before timing out.
    * @param {MergedRequestInit} [opts.fetchOptions] - Additional `RequestInit` options to be passed to `fetch` calls.
    * @param {Fetch} [opts.fetch] - Specify a custom `fetch` function implementation.
@@ -294,13 +252,13 @@ export class Knock {
    * @param {Record<string, string | undefined>} opts.defaultQuery - Default query parameters to include with every request to the API.
    */
   constructor({
-    baseURL = readEnv('KNOCK_BASE_URL'),
+    baseURL = readEnv('KNOCK_MGMT_BASE_URL'),
     serviceToken = readEnv('KNOCK_SERVICE_TOKEN'),
     ...opts
   }: ClientOptions = {}) {
     if (serviceToken === undefined) {
-      throw new Errors.KnockError(
-        "The KNOCK_SERVICE_TOKEN environment variable is missing or empty; either provide it, or instantiate the Knock client with an serviceToken option, like new Knock({ serviceToken: 'My Service Token' }).",
+      throw new Errors.KnockMgmtError(
+        "The KNOCK_SERVICE_TOKEN environment variable is missing or empty; either provide it, or instantiate the KnockMgmt client with an serviceToken option, like new KnockMgmt({ serviceToken: 'My Service Token' }).",
       );
     }
 
@@ -311,14 +269,14 @@ export class Knock {
     };
 
     this.baseURL = options.baseURL!;
-    this.timeout = options.timeout ?? Knock.DEFAULT_TIMEOUT /* 1 minute */;
+    this.timeout = options.timeout ?? KnockMgmt.DEFAULT_TIMEOUT /* 1 minute */;
     this.logger = options.logger ?? console;
     const defaultLogLevel = 'warn';
     // Set default logLevel early so that we can log a warning in parseLogLevel.
     this.logLevel = defaultLogLevel;
     this.logLevel =
       parseLogLevel(options.logLevel, 'ClientOptions.logLevel', this) ??
-      parseLogLevel(readEnv('KNOCK_LOG'), "process.env['KNOCK_LOG']", this) ??
+      parseLogLevel(readEnv('KNOCK_MGMT_LOG'), "process.env['KNOCK_MGMT_LOG']", this) ??
       defaultLogLevel;
     this.fetchOptions = options.fetchOptions;
     this.maxRetries = options.maxRetries ?? 2;
@@ -355,7 +313,7 @@ export class Knock {
         if (value === null) {
           return `${encodeURIComponent(key)}=`;
         }
-        throw new Errors.KnockError(
+        throw new Errors.KnockMgmtError(
           `Cannot stringify type ${typeof value}; Expected string, number, boolean, or null. If you need to pass nested query parameters, you can manually encode them, e.g. { query: { 'foo[key1]': value1, 'foo[key2]': value2 } }, and please open a GitHub issue requesting better support for your use case.`,
         );
       })
@@ -623,7 +581,7 @@ export class Knock {
     options: FinalRequestOptions,
   ): Pagination.PagePromise<PageClass, Item> {
     const request = this.makeRequest(options, null, undefined);
-    return new Pagination.PagePromise<PageClass, Item>(this as any as Knock, request, Page);
+    return new Pagination.PagePromise<PageClass, Item>(this as any as KnockMgmt, request, Page);
   }
 
   async fetchWithTimeout(
@@ -788,7 +746,7 @@ export class Knock {
         Accept: 'application/json',
         'User-Agent': this.getUserAgent(),
         'X-Stainless-Retry-Count': String(retryCount),
-        ...(options.timeout ? { 'X-Stainless-Timeout': String(options.timeout) } : {}),
+        ...(options.timeout ? { 'X-Stainless-Timeout': String(Math.trunc(options.timeout / 1000)) } : {}),
         ...getPlatformHeaders(),
       },
       this.authHeaders(options),
@@ -839,10 +797,10 @@ export class Knock {
     }
   }
 
-  static Knock = this;
+  static KnockMgmt = this;
   static DEFAULT_TIMEOUT = 60000; // 1 minute
 
-  static KnockError = Errors.KnockError;
+  static KnockMgmtError = Errors.KnockMgmtError;
   static APIError = Errors.APIError;
   static APIConnectionError = Errors.APIConnectionError;
   static APIConnectionTimeoutError = Errors.APIConnectionTimeoutError;
@@ -872,20 +830,20 @@ export class Knock {
   environments: API.Environments = new API.Environments(this);
   variables: API.Variables = new API.Variables(this);
 }
-Knock.Templates = Templates;
-Knock.EmailLayouts = EmailLayouts;
-Knock.Commits = Commits;
-Knock.Partials = Partials;
-Knock.Translations = Translations;
-Knock.Workflows = Workflows;
-Knock.MessageTypes = MessageTypes;
-Knock.Auth = Auth;
-Knock.APIKeys = APIKeys;
-Knock.ChannelGroups = ChannelGroups;
-Knock.Channels = Channels;
-Knock.Environments = Environments;
-Knock.Variables = Variables;
-export declare namespace Knock {
+KnockMgmt.Templates = Templates;
+KnockMgmt.EmailLayouts = EmailLayouts;
+KnockMgmt.Commits = Commits;
+KnockMgmt.Partials = Partials;
+KnockMgmt.Translations = Translations;
+KnockMgmt.Workflows = Workflows;
+KnockMgmt.MessageTypes = MessageTypes;
+KnockMgmt.Auth = Auth;
+KnockMgmt.APIKeys = APIKeys;
+KnockMgmt.ChannelGroups = ChannelGroups;
+KnockMgmt.Channels = Channels;
+KnockMgmt.Environments = Environments;
+KnockMgmt.Variables = Variables;
+export declare namespace KnockMgmt {
   export type RequestOptions = Opts.RequestOptions;
 
   export import EntriesCursor = Pagination.EntriesCursor;