+++ src/index.ts
@@ -1,13 +1,13 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-export { Knock as default } from './client';
+export { KnockMgmt as default } from './client';
 
-export { type Uploadable, toFile } from './uploads';
-export { APIPromise } from './api-promise';
-export { Knock, type ClientOptions } from './client';
-export { PagePromise } from './pagination';
+export { type Uploadable, toFile } from './core/uploads';
+export { APIPromise } from './core/api-promise';
+export { KnockMgmt, type ClientOptions } from './client';
+export { PagePromise } from './core/pagination';
 export {
-  KnockError,
+  KnockMgmtError,
   APIError,
   APIConnectionError,
   APIConnectionTimeoutError,
@@ -20,4 +20,4 @@ export {
   InternalServerError,
   PermissionDeniedError,
   UnprocessableEntityError,
-} from './error';
+} from './core/error';