+++ tests/api-resources/translations.test.ts
@@ -1,8 +1,8 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-import Knock from '@knocklabs/mgmt';
+import KnockMgmt from '@knocklabs/mgmt';
 
-const client = new Knock({
+const client = new KnockMgmt({
   serviceToken: 'My Service Token',
   baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
 });
@@ -35,7 +35,7 @@ describe('resource translations', () => {
         },
         { path: '/_stainless_unknown_path' },
       ),
-    ).rejects.toThrow(Knock.NotFoundError);
+    ).rejects.toThrow(KnockMgmt.NotFoundError);
   });
 
   // skipped: currently no good way to test endpoints defining callbacks, Prism mock server will fail trying to reach the provided callback url
@@ -68,7 +68,7 @@ describe('resource translations', () => {
         },
         { path: '/_stainless_unknown_path' },
       ),
-    ).rejects.toThrow(Knock.NotFoundError);
+    ).rejects.toThrow(KnockMgmt.NotFoundError);
   });
 
   // skipped: currently no good way to test endpoints defining callbacks, Prism mock server will fail trying to reach the provided callback url